Move sink provider to avoid extra deps
This moves the sink provider implementation to avoid requiring people
who consume metrics-server as a library to pull in extra dependencies.
